requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
bcrypt==4.1.1
PyJWT==2.8.0
python-jose[cryptography]==3.3.0